dev: 构建更新

+++ bin/index.js
@@ -12,26 +12,29 @@ programRegister(program,{
     // 创建项目
     create: async (projectName) => {
         showMain() // 显示文字
+
+        const hasUpgrade = await updateChk() // 检查更新
+        if(hasUpgrade) return // 拦截
+
+
         const { repo } =  await create(config) // form信息
         const path = `./${projectName}`
         if(directoryExists(path)){ // 同名目录
             const res = await confirmRemake() // 是否删除
-            if(res) await deleteFile(path)
+            if(res) await deleteFile(path) // 删除目录
+            else return
         }
-        const res = await confirmCreate()
+
+        const res = await confirmCreate() // 确认创建
         if(res) { // 校验参数, 准备下载
             const res = await download({projectName,repo})
             showDirTree(res)
         }
+
     },
     // 查看目录结构
-    dir: async (dir) => {
-        if(!dir) dir = './'
-        showDirTree(dir)
-    },
+    dir: (dir) => showDirTree(dir),
     // 手动更新项目
-    upgrade: async () => {
-        updateChk()
-    }
+    upgrade: () => updateChk(true)
 })
